[PORT FROM MAIN] Save the export progress when rotating the phone
BZ: 37621

When rotating the phone which is exporting movies,
the export progress will be lost. The progress bar
will appear to be 0 for a few seconds. In this patch,
The export progress is saved when the video editor
activity is destroyed.

Category: aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: Yes

Change-Id: I5a3aeb43de547db5bbb6458e3d9e6ade26acb3cd
Orig-Change-Id: Idb7c201f110ce5738b0963f9be42952c87db811f
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>