fix the javacrash after unlocking phone in landscape mode
BZ: 80630

Root cause: in this test case, when user unlock the phone in the landscape
the view of phone will stay in the portrait with a short time, then the view
will go to landscape, and the code will get width of the portrait view but
use it in the landscape. It's android original design, we add a parameter
check to avoid crash.

Change-Id: I476f167653bf336d768c37851d4e5f21d6ed77fa
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/87436
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>