[PORT FROM MAIN][Movie studio]: fix the javacrash after unlocking phone in landscape mode
BZ: 80630

Root cause: in this test case, when user unlock the phone in the landscape
the view of phone will stay in the portrait with a short time, then the view
will go to landscape, and the code will get width of the portrait view but
use it in the landscape. It's android original design, we add a parameter
check to avoid crash.

Change-Id: Ie21bbcdcb7e74a64e5c29d91582fa47afc3f49d3
Orig-Change-Id: I0ebab13062a5e189484d17d4072f00a981df1c33
Category: aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Signed-off-by: chuzhoux <chux.zhou@intel.com>