[PORT FROM MAIN][Movie studio]: improve the UE when tap 'PREV' continuous during video playback
BZ: 90529

it's an android defect, the reference phone(i9250) also has this issue
root cause:
there is a message queue in preview thread shared by three message producers(
startPreviewPlayback,previewFrame,renderMediaItemFrame). In the test case of
bug 90529, many prev click operations will invoke many messages from "PreviewFrame",
according to the android design, adding these messages will remove the message
produced by "startPreviewPlayback", but the preview thread is waitting the
result of this message, and it doesn't know this message removed, so the issue happens.
This patch skips the prev click when preview thread is waitting the result of
the message from "startPreviewPlayback" to improve the UE in this test case.
This issue will happen when the interval between two "prev click" operation
is small and many "prev click" operations in a short time, so normal operation
won't be impacted by this patch.

Change-Id: I9876c4328bf2b1ca32e205684cc89bcdd17897bc
Orig-Change-Id: Ia33ee539ea05c56740175fdc6fd4e1b58b7bb2d0
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/121764
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>