[PORT FROM MAIN][Movie studio]: improve the UX when edit the saved title template
BZ: 100462

There are two issues are improved in this patch.
1. there is a saved title in the project, if rotate the phone when edit the
saved title(choose a new title template), but the new picked title template
is not updated after rotation.
2. at the end of issue 1, the user has to pick the new title template after
rotate the phone if he wants to change the title template, then he saves the
title, but the view of title in the thumbnail area is not updated.

Change-Id: I83b0f619202944f892246f7b6e4b63dc3fcec04a
Orig-Change-Id: I925d388cc5c0ee473b5fd76030fca15cfea25b46
Category: aosp improvement
Domain: Video.Media-video editor
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>