[Movie studio]: improve the UX when edit the saved title template
BZ: 100462

There are two issues are improved in this patch.
1. there is a saved title in the project, if rotate the phone when edit the
saved title(choose a new title template), but the new picked title template
is not updated after rotation.
2. at the end of issue 1, the user has to pick the new title template after
rotate the phone if he wants to change the title template, then he saves the
title, but the view of title in the thumbnail area is not updated.

Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>

Change-Id: I925d388cc5c0ee473b5fd76030fca15cfea25b46
Reviewed-on: http://android.intel.com:8080/126056
Reviewed-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>